pyproject.toml
[tool.poetry]
name = "django-nginx-gunicorn-deploy-using-docker-for-aws"
version = "0.1.0"
description = ""
authors = ["Andreas Cormack <cormackandy@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
gunicorn = "^22.0.0"
djangorestframework = "^3.15.2"
Markdown = "^3.6"
django-filter = "^24.2"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
daphne = "^4.1.2"
channels-redis = "^4.2.0"
boto3 = "^1.18.67"
google-auth = "^2.3.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
